+++ tests/test-silent-publish.php
@@ -129,6 +129,41 @@ class Silent_Publish_Test extends WP_UnitTestCase {
 		$this->assertNotFalse( has_action( 'publish_post', array( 'c2c_SilentPublish', 'publish_post' ), 1, 1 ) );
 	}
 
+	/*
+	 * get_post_types()
+	 */
+
+	public function test_get_post_types( $extra_post_types = array() ) {
+		$expected = array_merge( array( 'post', 'page' ), $extra_post_types );
+
+		$this->assertEquals( $expected, c2c_SilentPublish::get_post_types() );
+	}
+
+	public function test_c2c_silent_publish_post_types_with_custom_post_type() {
+		register_post_type( 'private', array( 'public' => false, 'name' => 'Private' ) );
+		register_post_type( 'book', array( 'public' => true, 'name' => 'Book' )	);
+
+		$this->test_get_post_types( array( 'book' ) );
+	}
+
+	/*
+	 * filter: c2c_silent_publish_post_types
+	 */
+
+	 public function test_filter_c2c_silent_publish_post_types() {
+		register_post_type( 'private', array( 'public' => false, 'name' => 'Private' ) );
+		register_post_type( 'book', array( 'public' => true, 'name' => 'Book' )	);
+
+		add_filter( 'c2c_silent_publish_post_types', function ( $p ) {
+			$p[] = 'private';
+			$p = array_flip( $p );
+			unset( $p[ 'book' ] );
+			return array_flip( $p );
+		} );
+
+		$this->test_get_post_types( array_keys( 'private' ) );
+	}
+
 	public function test_non_silently_published_post_publishes_without_silencing() {
 		$post_id = $this->factory->post->create( array( 'post_status' => 'draft' ) );